pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srellm"
version = "0.1.0"
description = "SRE LLM - A package for SRE tasks with LLM integration"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "qdrant_client",
    "uvicorn",
    "pydantic",
    "faiss-cpu",
    "sentence-transformers",
    "numpy",
    "pandas",
    "tqdm",
    "ipython",
    "ipdb",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["srellm*", "rag*", "serving*", "tools*"]
exclude = ["data*", "evals*", "troubleshoot*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true